Google wallet and terminal implementation.
Google wallet and terminal implementation.
In this PR, the following changes were brought.

- Ability to add Google wallet pass.
- Ability to scan rotating barcode at terminal.
- Author tag on top of backend files.

Note that the validation made when scanning the QR code isn't complete. Normally, we should validate the ticket and membership amount and status, make sure that it hasn't been too long since the qr code was generated, and we should generate a totp using a hex key and `hmacsha1` and compare it with the one in the QR code. Though, I wasn't able to do this. See `ScannerService.java`.